Initial version of validation.py
I'm currently running this as a stand-alone program (i.e. python validation.py) rather than integrating it in the main program. It should be producing results similar to that of test_script_ud.py (with some additional labeling and format changes) but I haven't confirmed this yet; also it is just checking the first six records but this can easily be changed. Output will be in Validations_out1.txt. Also this uses the file data/text/PETR.UnitTest.records.parsing_parsed_udpipe_2.1.xml